% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cali_DDS}
\alias{cali_DDS}
\title{This function implements a calibration algorithm based on the Direct Search method.}
\usage{
cali_DDS(
  fitness,
  lst_OtherData,
  x_Min,
  x_Max,
  x_Init = NULL,
  max_iter = 100L,
  r = 0.2
)
}
\arguments{
\item{fitness}{A function handle to the objective function that returns a numeric value}

\item{lst_OtherData}{A list of other data that may be passed to the fitness function}

\item{x_Min}{A NumericVector of lower bounds for each parameter}

\item{x_Max}{A NumericVector of upper bounds for each parameter}

\item{x_Init}{An optional initial solution (NumericVector), if NULL, the midpoint of `x_Min` and `x_Max` is used}

\item{max_iter}{The maximum number of iterations to perform (default is 100)}

\item{r}{The perturbation factor (default is 0.2)}
}
\value{
A List containing:
  - `x_Best`: The best parameter set found during the search
  - `y_Best`: The corresponding objective value of the best parameter set
}
\description{
It attempts to find the optimal parameter values that minimize the given objective (fitness) function.
}
