% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{eva}
\alias{eva}
\alias{eva_NSE}
\alias{eva_KGE}
\title{Evalute matrics}
\usage{
eva_NSE(num_Sim, num_Obs)

eva_KGE(num_Sim, num_Obs, factor_r = 1, factor_alpha = 1, factor_beta = 1)
}
\arguments{
\item{num_Sim}{A numeric vector of simulated values.}

\item{num_Obs}{A numeric vector of observed values. NA values are removed along with corresponding values in num_Sim.}

\item{factor_r, factor_alpha, factor_beta}{A double specifying the weight for the correlation term (r - 1), (alpha - 1) and (beta - 1). Default is 1.0.}
}
\value{
A double representing the Evalute matrics.
}
\description{
Evalute matrics
}
