% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_inflow_cells}
\alias{get_inflow_cells}
\alias{get_inflow_lastcell}
\alias{get_step_param}
\title{Get Inflow Cells}
\usage{
get_inflow_cells(int_Outflow)

get_inflow_lastcell(int_Outflow)

get_step_param(int_Outflow)
}
\arguments{
\item{int_Outflow}{(vector of int) The cell number of the next cell. The cell number must range from 1 to the length of the cells.
If the cell has no outflow, the number should be set to itself.}
}
\value{
A list where each element is an IntegerVector containing the inflow cells for each cell.

A NumericMatrix where each row corresponds to a cell, and each column represents an inflow cell.

A list containing the step cells and last cell matrices.
}
\description{
This function calculates inflow cells based on the outflow vector.

This function creates a matrix of inflow cells for each cell based on the outflow vector.

This function returns a list of step cells and the corresponding last cell matrices.
}
