% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{unf}
\alias{unf}
\alias{read_unf}
\alias{write_unf}
\title{read_UNF - Read data from UNF file}
\usage{
read_unf(fn_UNF)

write_unf(data_Export, fn_UNF)
}
\arguments{
\item{fn_UNF}{A string, name of the UNF file}

\item{data_Export}{A numeric or integer vector, data to be exported}
}
\value{
A numeric or integer vector by reading the UNF file
}
\description{
read_UNF - Read data from UNF file

write_UNF - Write data to UNF file
}
