% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{utilis}
\alias{utilis}
\alias{subset_get}
\alias{subset_get_logical}
\alias{subset_put}
\alias{subset_add}
\title{**utilize functions**}
\usage{
subset_get(vec_Data, int_Index)

subset_get_logical(vec_Data, int_Index)

subset_put(vec_Data, int_Index, vec_DataPut)

subset_add(vec_Data, int_Index, vec_DataAdd)
}
\arguments{
\item{vec_Data}{(vector of num / int) data}

\item{int_Index}{(vector of int) index of subset}

\item{vec_DataPut}{(vector of num / int) data, to refresh}

\item{vec_DataAdd}{(vector of num / int) data, to refresh}
}
\description{
**utilize functions**
}
