% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{WaterGAP3}
\alias{WaterGAP3}
\alias{WaterGAP3_N}
\title{WaterGAP3}
\usage{
WaterGAP3_N(
  name_Region,
  mark_Time,
  n_time,
  n_spat,
  ATMOS_precipitation_mm,
  ATMOS_temperature_Cel,
  ATMOS_solarRadiat_MJ,
  ATMOS_solarRadiatClearSky_MJ,
  Upstream_cellNumber_int,
  Upstream_streamflow_m3,
  SNOW_ice_mm,
  LAND_area_km2,
  LAND_albedo_1,
  LAND_snowAlbedo_1,
  LAND_builtRatio_1,
  LAND_interceptWater_mm,
  LAND_interceptCapacity_mm,
  SOIL_water_mm,
  SOIL_capacity_mm,
  SOIL_potentialPercola_mm,
  GROUND_water_mm,
  RIVER_water_m3,
  RIVER_length_km,
  RIVER_velocity_km,
  CELL_elevation_m,
  CELL_cellNumberStep_int,
  CELL_inflowCellNumberStep_int,
  Lake_cellNumber_int,
  Lake_water_m3,
  Lake_area_km2,
  Lake_capacity_m3,
  Lake_albedo_1,
  Riverlak_cellNumber_int,
  Riverlak_water_m3,
  Riverlak_area_km2,
  Riverlak_capacity_m3,
  Riverlak_albedo_1,
  param_ATMOS_thr_Ts,
  param_SNOW_fac_f,
  param_SNOW_fac_Tmelt,
  param_EVATRANS_prt_alpha,
  param_EVATRANS_vic_gamma,
  param_EVATRANS_sup_k,
  param_EVATRANS_sup_gamma,
  param_EVATRANS_wat_petmax,
  param_INFILT_hbv_beta,
  param_PERCOLA_wat_01,
  param_PERCOLA_wat_k,
  param_PERCOLA_wat_thresh,
  param_BASEFLOW_sur_k,
  param_Lake_acp_storeFactor,
  param_Lake_acp_gamma,
  param_Riverlak_lin_storeFactor,
  path_FinalState = "NonExport",
  path_VariExport = "NonExport"
)
}
\arguments{
\item{name_Region}{A short identifier of continents: "eu", "af, "as", "au", "na", "sa".}

\item{mark_Time}{A short identifier of time.}

\item{n_time}{(int) number of time steps}

\item{n_spat}{(int) number of spatial units}

\item{ATMOS_precipitation_mm}{(mm/m2/TS) precipitaion volum}

\item{ATMOS_temperature_Cel}{(Cel) the average air temperature in the time phase}

\item{ATMOS_solarRadiat_MJ}{(MJ/m2/TS) the solar radiation that actually reaches the earths surface}

\item{Upstream_cellNumber_int}{Upstream cell number}

\item{Upstream_streamflow_m3}{Upstream}

\item{SNOW_ice_mm}{(mm/m2) water equivalent of **ice** in snowpack}

\item{LAND_albedo_1}{<0, 1> albedo of the region}

\item{LAND_interceptWater_mm}{(mm/m2) initial water volume that can be **intercepted**}

\item{LAND_interceptCapacity_mm}{(mm/m2) average intercept Capacity (maximal storage capacity)}

\item{SOIL_water_mm}{(mm/m2) water volume in `soilLy`}

\item{SOIL_capacity_mm}{(mm/m2) average soil Capacity (maximal storage capacity)}

\item{SOIL_potentialPercola_mm}{<0.01, 7> (mm/m2/TS) **potential** percolation}

\item{GROUND_water_mm}{(mm/m2/TS) water volume in `groundLy`}

\item{RIVER_water_m3}{(m3) river water volume}

\item{RIVER_length_km}{(km) river length}

\item{RIVER_velocity_km}{(km/TS) river velocity}

\item{CELL_cellNumberStep_int}{(int) step number of basin cells}

\item{CELL_inflowCellNumberStep_int}{(int) step number of inflow cells in the basin}

\item{Lake_cellNumber_int}{(int) lake cell number}

\item{Lake_water_m3}{(m3) lake water volume}

\item{Lake_area_km2}{(km2) area of the lake}

\item{Lake_capacity_m3}{(m3) capacity of the lake}

\item{Riverlak_cellNumber_int}{(int) river-lake cell number}

\item{Riverlak_water_m3}{(m3) water volume in the river-lake system}

\item{Riverlak_area_km2}{(km2) area of the river-lake system}

\item{Riverlak_capacity_m3}{(m3) total capacity of the river-lake system}

\item{path_FinalState}{Directory path where final state variables will be saved. If "NonExport", final states will not be saved.}

\item{path_VariExport}{Directory path where model output variables (e.g., soil moisture, runoff) will be saved. If "NonExport", no variables will be exported.}
}
\value{
streamflow m3
}
\description{
WaterGAP3
}
