% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{WaterGAP3_H}
\alias{WaterGAP3_H}
\title{WaterGAP3}
\usage{
WaterGAP3_H(
  n_time,
  n_spat,
  AtmoS_precipitation_mm,
  AtmoS_temperature_Cel,
  AtmoS_solarRadiat_MJ,
  snoW_ice_mm,
  soiL_water_mm,
  soiL_capacity_mm,
  soiL_potentialPercola_mm,
  grounD_water_mm,
  grounD_capacity_mm,
  riveR_water_m3,
  riveR_length_km,
  riveR_velocity_km,
  celL_landArea_km2,
  celL_cellNumberStep_int,
  celL_inflowCellNumberStep_int,
  param_atmos_thr_Ts,
  param_snow_fac_f,
  param_snow_fac_Tmelt,
  param_evatrans_tur_k,
  param_evatrans_ubc_gamma,
  param_infilt_hbv_beta,
  param_percola_arn_k,
  param_percola_arn_thresh,
  param_baseflow_grf_gamma,
  if_allVariExport = FALSE
)
}
\arguments{
\item{n_time}{(int) number of time steps}

\item{n_spat}{(int) number of spatial units}

\item{AtmoS_precipitation_mm}{(mm/m2/TS) precipitaion volum}

\item{AtmoS_temperature_Cel}{(Cel) the average air temperature in the time phase}

\item{AtmoS_solarRadiat_MJ}{(MJ/m2/TS) the solar radiation that actually reaches the earths surface}

\item{snoW_ice_mm}{(mm/m2) water equivalent of **ice** in snowpack}

\item{soiL_water_mm}{(mm/m2) water volume in `soilLy`}

\item{soiL_capacity_mm}{(mm/m2) average soil Capacity (maximal storage capacity)}

\item{soiL_potentialPercola_mm}{<0.01, 7> (mm/m2/TS) **potential** percolation}

\item{grounD_water_mm}{(mm/m2/TS) water volume in `groundLy`}

\item{grounD_capacity_mm}{(mm/m2) water storage capacity in `groundLy`}

\item{riveR_water_m3}{(m3) river water volume}

\item{riveR_length_km}{(km) river length}

\item{riveR_velocity_km}{(km/TS) river velocity}

\item{celL_landArea_km2}{(km2) land area of the basin}

\item{celL_cellNumberStep_int}{(int) step number of basin cells}

\item{celL_inflowCellNumberStep_int}{(int) step number of inflow cells in the basin}

\item{param_atmos_thr_Ts}{<-1, 3> (Cel) threshold air temperature that snow, parameter for [atmosSnow_ThresholdT()]}

\item{param_snow_fac_f}{<0.05, 2> (mm/m2/h/Cel) potential melt volum per Cel per hour parameter for [snowMelt_Factor()]}

\item{param_snow_fac_Tmelt}{<0, 3> (Cel) snow melt temperature parameter for [snowMelt_Factor()]}

\item{param_evatrans_tur_k}{<0.6, 1> parameter for [evatransPotential_TurcWendling()], higher value when closer to the sea}

\item{param_evatrans_ubc_gamma}{<0.5, 2> parameter for [evatransActual_UBC()]}

\item{param_infilt_hbv_beta}{<0.001, 5> parameters for [infilt_HBV()]}

\item{param_percola_arn_k}{<0.1, 1> exponential parameter for [percola_ThreshPow()]}

\item{param_percola_arn_thresh}{<0.1, 0.9> coefficient parameter for [percola_ThreshPow()]}

\item{param_baseflow_grf_gamma}{<2, 7> exponential parameter for [baseflow_GR4Jfix()]}

\item{if_allVariExport}{(bool) whether export all the state variables (default in false)}
}
\value{
streamflow m3
}
\description{
WaterGAP3
}
