% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{WaterGAP3_HLR}
\alias{WaterGAP3_HLR}
\title{WaterGAP3}
\usage{
WaterGAP3_HLR(
  n_time,
  n_spat,
  ATMOS_precipitation_mm,
  ATMOS_temperature_Cel,
  ATMOS_potentialEvatrans_mm,
  SNOW_ice_mm,
  SOIL_water_mm,
  SOIL_capacity_mm,
  SOIL_potentialPercola_mm,
  GROUND_water_mm,
  GROUND_capacity_mm,
  RIVER_water_m3,
  RIVER_length_km,
  RIVER_velocity_km,
  CELL_landArea_km2,
  CELL_cellNumberStep_int,
  CELL_inflowCellNumberStep_int,
  Lake_cellNumber_int,
  Lake_water_m3,
  Lake_area_km2,
  Lake_capacity_m3,
  Riverlak_cellNumber_int,
  Riverlak_water_m3,
  Riverlak_area_km2,
  Riverlak_capacity_m3,
  Reservoi_cellNumber_int,
  Reservoi_water_m3,
  Reservoi_area_km2,
  Reservoi_capacity_m3,
  Reservoi_demand_m3,
  Reservoi_yearInflow_m3,
  Reservoi_yearDemand_m3,
  Reservoi_yearRelase_m3,
  Reservoi_isOperateStart_01,
  Reservoi_isIrrigate_01,
  param_ATMOS_thr_Ts,
  param_SNOW_fac_f,
  param_SNOW_fac_Tmelt,
  param_EVATRANS_ubc_gamma,
  param_INFILT_hbv_beta,
  param_PERCOLA_arn_k,
  param_PERCOLA_arn_thresh,
  param_BASEFLOW_grf_gamma,
  param_Evalake_vic_gamma,
  param_Lake_acp_storeFactor,
  param_Lake_acp_gamma,
  param_Riverlak_lin_storeFactor,
  param_Reservoi_han_alpha,
  param_Reservoi_han_kDemand,
  if_allVariExport = FALSE
)
}
\arguments{
\item{n_time}{(int) number of time steps}

\item{n_spat}{(int) number of spatial units}

\item{ATMOS_precipitation_mm}{(mm/m2/TS) precipitaion volum}

\item{ATMOS_temperature_Cel}{(Cel) the average air temperature in the time phase}

\item{ATMOS_potentialEvatrans_mm}{(mm/m2/TS) **potential / reference** evapotranspiration}

\item{SNOW_ice_mm}{(mm/m2) water equivalent of **ice** in snowpack}

\item{SOIL_water_mm}{(mm/m2) water volume in `soilLy`}

\item{SOIL_capacity_mm}{(mm/m2) average soil Capacity (maximal storage capacity)}

\item{SOIL_potentialPercola_mm}{<0.01, 7> (mm/m2/TS) **potential** percolation}

\item{GROUND_water_mm}{(mm/m2/TS) water volume in `groundLy`}

\item{GROUND_capacity_mm}{(mm/m2) water storage capacity in `groundLy`}

\item{RIVER_water_m3}{(m3) river water volume}

\item{RIVER_length_km}{(km) river length}

\item{RIVER_velocity_km}{(km/TS) river velocity}

\item{CELL_landArea_km2}{(km2) land area of the basin}

\item{CELL_cellNumberStep_int}{(int) step number of basin cells}

\item{CELL_inflowCellNumberStep_int}{(int) step number of inflow cells in the basin}

\item{Lake_cellNumber_int}{(int) lake cell number}

\item{Lake_water_m3}{(m3) lake water volume}

\item{Lake_area_km2}{(km2) area of the lake}

\item{Lake_capacity_m3}{(m3) capacity of the lake}

\item{Riverlak_cellNumber_int}{(int) river-lake cell number}

\item{Riverlak_water_m3}{(m3) water volume in the river-lake system}

\item{Riverlak_area_km2}{(km2) area of the river-lake system}

\item{Riverlak_capacity_m3}{(m3) total capacity of the river-lake system}

\item{Reservoi_cellNumber_int}{(int) reservoir cell number}

\item{Reservoi_water_m3}{(m3) volume of water in the reservoir}

\item{Reservoi_area_km2}{(km2) surface area of the reservoir}

\item{Reservoi_capacity_m3}{(m3) reservoir capacity}

\item{Reservoi_demand_m3}{(m3/TS) water demand from the reservoir}

\item{Reservoi_yearInflow_m3}{(m3/TS) annual mean inflow to the reservoir}

\item{Reservoi_yearDemand_m3}{(m3/TS) annual mean water demand from the reservoir}

\item{Reservoi_yearRelase_m3}{(m3/TS) reference to annual mean water release from the reservoir}

\item{Reservoi_isOperateStart_01}{(01) indicates if reservoir operation has started (0 or 1)}

\item{Reservoi_isIrrigate_01}{(01) indicates if the reservoir is used for irrigation (0 or 1)}

\item{param_ATMOS_thr_Ts}{<-1, 3> (Cel) threshold air temperature that snow, parameter for [atmosSnow_ThresholdT()]}

\item{param_SNOW_fac_f}{<0.05, 2> (mm/m2/h/Cel) potential melt volum per Cel per hour parameter for [snowMelt_Factor()]}

\item{param_SNOW_fac_Tmelt}{<0, 3> (Cel) snow melt temperature parameter for [snowMelt_Factor()]}

\item{param_EVATRANS_ubc_gamma}{<0.5, 2> parameter for [evatransActual_UBC()]}

\item{param_INFILT_hbv_beta}{<0.001, 5> parameters for [infilt_HBV()]}

\item{param_PERCOLA_arn_k}{<0.1, 1> exponential parameter for [percola_ThreshPow()]}

\item{param_PERCOLA_arn_thresh}{<0.1, 0.9> coefficient parameter for [percola_ThreshPow()]}

\item{param_BASEFLOW_grf_gamma}{<2, 7> exponential parameter for [baseflow_GR4Jfix()]}

\item{param_Evalake_vic_gamma}{same as param_EVATRANS_vic_gamma}

\item{param_Lake_acp_storeFactor}{<uknow> parameter for [lake_AcceptPow()],}

\item{param_Lake_acp_gamma}{<uknow> parameter for [lake_AcceptPow()],}

\item{param_Riverlak_lin_storeFactor}{<uknow> parameter for [riverlak_LinearResorvoir()],}

\item{param_Reservoi_han_alpha}{<uknow> parameter for [reservoi_Hanasaki()],}

\item{param_Reservoi_han_kDemand}{<uknow> parameter for [reservoi_Hanasaki()],}

\item{if_allVariExport}{(bool) whether export all the state variables (default in false)}
}
\value{
streamflow m3
}
\description{
WaterGAP3
}
