% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{save_wgmat}
\alias{save_wgmat}
\alias{wgmat}
\alias{load_wgmat}
\alias{bind_wgmat}
\title{Save a NumericMatrix to a binary file}
\usage{
save_wgmat(matrix, filename)

load_wgmat(filename)

bind_wgmat(input_files, output_file)
}
\arguments{
\item{matrix}{The NumericMatrix to be saved.}

\item{filename}{The name of the binary file to load the matrix from.}

\item{input_files}{A vector of file paths to the input binary files.}

\item{output_file}{The file path to save the combined matrix.}
}
\value{
A NumericMatrix containing the data from the binary file.
}
\description{
This function saves a NumericMatrix to a binary file, with the matrix
data stored in a row-major order (transposed).

This function loads a NumericMatrix from a binary file that was saved
using the `save_wgmat` function. The matrix is reconstructed in R's
column-major order.

This function binds several matrices from binary files into a single
matrix and saves it to a new binary file. All input matrices must have
the same number of columns. The function validates the dimensions of
the matrices before combining them.
}
