% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_variable.R
\name{all_vari}
\alias{all_vari}
\title{variable define}
\arguments{
\item{n_time}{(int) number of time steps}

\item{n_spat}{(int) number of spatial units}

\item{Time_dayOfYear_, Time_dayOfYear_d}{<1, 366> (d) the number of the day in the year between 1 (1 January) and 365 or 366 (31 December)}

\item{ATMOS_precipitation_mm}{(mm/m2/TS) precipitaion volum}

\item{ATMOS_rain_mm}{(mm/m2/TS) precipitation in rain form}

\item{ATMOS_SNOW_mm}{(mm/m2/TS) precipitation in snow form}

\item{ATMOS_temperature_Cel}{(Cel) the average air temperature in the time phase}

\item{ATMOS_temperatureMax_Cel}{(Cel) the maximal air temperature in the time phase}

\item{ATMOS_temperatureMin_Cel}{(Cel) the minimal air temperature in the time phase}

\item{ATMOS_solarRadiat_MJ}{(MJ/m2/TS) the solar radiation that actually reaches the earths surface}

\item{ATMOS_netRadiat_MJ}{(MJ/m2/TS) the balance between the energy absorbed, reflected and emitted by the earths surface or the difference between the incoming net shortwave (Rns) and the net outgoing longwave (Rnl) radiation}

\item{ATMOS_relativeHumidity_1}{(0, 1) relative humidity}

\item{ATMOS_vaporPress_hPa}{(hPa) actual vapour pressure, can be calculated by [ATMOS_VaporPress()]}

\item{ATMOS_saturatVaporPress_hPa}{(hPa) saturation vapour pressure at `ATMOS_temperature_Cel`, can be calculated by [ATMOS_SaturatVaporPress()]}

\item{ATMOS_windSpeed_m_s}{(m/s) measured wind speed at z m above ground surface}

\item{ATMOS_windMeasureHeight_m}{(m) height of measurement above ground surface}

\item{ATMOS_windSpeed2m_m_s}{(m/s) wind speed at 2 m above ground surface}

\item{ATMOS_potentialEvatrans_mm}{(mm/m2/TS) **potential / reference** evapotranspiration}

\item{LAND_albedo_1}{<0, 1> albedo of the region}

\item{LAND_growUpDay_d}{growup day}

\item{LAND_latitude_Degree, CELL_latitude_deg}{(degree) average latitude}

\item{LAND_elevation_m}{(m) average elevation}

\item{LAND_impermeableFrac_1}{<0, 1> the maximum impermeable fraction when th soil is fully saturated}

\item{LAND_water_mm}{(mm/m2) **pounded water** volume in `landLy` and there is no limit, different than `LAND_interceptWater_mm`}

\item{LAND_interceptWater_mm}{(mm/m2) initial water volume that can be **intercepted**}

\item{LAND_interceptCapacity_mm}{(mm/m2) average intercept Capacity (maximal storage capacity)}

\item{LAND_actualEvatrans_mm}{(mm/m2/TS) **actual** evapotranspiration from `landLy`}

\item{LAND_infiltrat_mm}{(mm/m2/TS) infiltration from `landLy` to `soilLy`}

\item{LAND_leafAreaIndexMin_, LAND_leafAreaIndexMax_}{minimal and maximal LAI}

\item{SNOW_ice_mm}{(mm/m2) water equivalent of **ice** in snowpack}

\item{SOIL_fieldCapacityPerc_1}{<0, 1> the relative ratio that the water content can drainage by gravity}

\item{SOIL_water_mm}{(mm/m2) water volume in `soilLy`}

\item{SOIL_capacity_mm}{(mm/m2) average soil Capacity (maximal storage capacity)}

\item{SOIL_interflow_mm}{(mm/m2/TS) subsurface flow directly to the river}

\item{SOIL_actualEvatrans_mm}{(mm/m2/TS) **actual** evapotranspiration from `soilLy`}

\item{SOIL_potentialPercola_mm}{<0.01, 7> (mm/m2/TS) **potential** percolation}

\item{SOIL_potentialInteflow_mm}{<0.01, 7> (mm/m2/TS) **potential** interflow}

\item{SOIL_potentialCapirise_mm}{<0.01, 7> (mm/m2/TS) **potential** capillary rise}

\item{GROUND_water_mm}{(mm/m2/TS) water volume in `groundLy`}

\item{GROUND_capacity_mm}{(mm/m2) water storage capacity in `groundLy`}

\item{GROUND_lateral_mm}{(mm/m2/TS) lateral flow, exchange with outside region. It can be **NEGATIV**}

\item{GROUND_capillarise_mm}{(mm/m2/TS) capillary rise from `groundLy` to `soilLy`}

\item{GROUND_potentialLateral_mm}{<-7, 7> (mm/m2/TS) **potential** lateral flow}

\item{GROUND_potentialBaseflow_mm}{<0.01, 7> (mm/m2/TS) **potential** baseflow}

\item{CONFLUEN_cellInflow_m3}{(m3 / TS) input water volum in every routeline}

\item{CONFLUEN_inputWater_mm, LAND_runoff_mm, GROUND_baseflow_mm}{(mm/m2) input water volum in every routeline}

\item{CONFLUEN_iuh_1, CONFLUEN_iuhLand_1, CONFLUEN_iuhSoil_1, CONFLUEN_iuhGround_1}{(vector of num, sume() = 1) the ratio in every timestep, can be calculated by [confluenIUH_GR4J1()], [confluenIUH_GR4J2()]}

\item{CONFLUEN_responseTime_TS}{(TS) response or concentration time in every routeline}

\item{water_mm}{(mm/m2/TS) water volume in `soilLy`, `groundLy` or intercept of `landLy`, same as `SOIL_water_mm`, `GROUND_water_mm` or `LAND_interceptWater_mm`}

\item{capacity_mm}{(mm/m2) water storage capacity in `soilLy`, `groundLy` or intercept of `landLy`, same as `SOIL_capacity_mm`, `GROUND_capacity_mm` or`LAND_interceptCapacity_mm`}

\item{Lake_cellNumber_int}{(int) lake cell number}

\item{Lake_water_m3}{(m3) lake water volume}

\item{Lake_EVATRANS_mm}{(mm) evapotranspiration from the lake}

\item{Lake_capacity_m3}{(m3) capacity of the lake}

\item{Lake_area_km2}{(km2) area of the lake}

\item{Lake_inflow_m3}{(m3/TS) inflow of the lake}

\item{RIVER_water_m3}{(m3) river water volume}

\item{RIVER_inflow_m3}{(m3/TS) river inflow volume}

\item{RIVER_outflow_m3}{(m3/TS) river outflow volume}

\item{RIVER_velocity_km}{(km/TS) river velocity}

\item{RIVER_length_km}{(km) river length}

\item{Riverlak_cellNumber_int}{(int) river-lake cell number}

\item{Riverlak_area_km2}{(km2) area of the river-lake system}

\item{Riverlak_capacity_m3}{(m3) total capacity of the river-lake system}

\item{Riverlak_inflow_m3}{(m3) inflow volume to the river-lake system}

\item{Riverlak_water_m3}{(m3) water volume in the river-lake system}

\item{RIVER_upstreamInflow_m3}{(m3) Upstream flow}

\item{Reservoi_cellNumber_int}{(int) reservoir cell number}

\item{Reservoi_water_m3}{(m3) volume of water in the reservoir}

\item{Reservoi_capacity_m3}{(m3) reservoir capacity}

\item{Reservoi_EVATRANS_mm}{(mm) evapotranspiration from the reservoir}

\item{Reservoi_inflow_m3}{(m3/TS) inflow to the reservoir}

\item{Reservoi_demand_m3}{(m3/TS) water demand from the reservoir}

\item{Reservoi_potentialVolume_m3}{(m3) potential capacity of the reservoir}

\item{Reservoi_area_km2}{(km2) surface area of the reservoir}

\item{Reservoi_yearInflow_m3}{(m3/TS) annual mean inflow to the reservoir}

\item{Reservoi_yearDemand_m3}{(m3/TS) annual mean water demand from the reservoir}

\item{Reservoi_yearRelase_m3}{(m3/TS) reference to annual mean water release from the reservoir}

\item{Reservoi_isOperateStart_01}{(01) indicates if reservoir operation has started (0 or 1)}

\item{Reservoi_isIrrigate_01}{(01) indicates if the reservoir is used for irrigation (0 or 1)}

\item{CELL_landArea_km2}{(km2) land area of the basin}

\item{CELL_cellNumberStep_int}{(int) step number of basin cells}

\item{CELL_inflowCellNumberStep_int}{(int) step number of inflow cells in the basin}

\item{Upstream_cellNumber_int}{Upstream cell number}

\item{Upstream_streamflow_m3}{Upstream}

\item{if_allVariExport}{(bool) whether export all the state variables (default in false)}
}
\description{
In this topic all of the variable in the EDCHM will be defined.
}
