% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nc}
\alias{nc}
\alias{write_nc_WG3}
\alias{read_nc_WG3}
\alias{read_nc_dim_WG3}
\title{Write data to a NetCDF file in WaterGAP3 format}
\usage{
write_nc_WG3(
  mat_Data_WG3,
  dim_Time,
  dim_Spat,
  path_File,
  name_Variable,
  str_Continent,
  suffix_File
)

read_nc_WG3(path_File, name_Variable)

read_nc_dim_WG3(path_File, dim_name)
}
\arguments{
\item{mat_Data_WG3}{A numeric matrix containing the data to be written. Rows represent time steps, columns represent spatial units.}

\item{dim_Time}{An integer vector specifying the time dimension values.}

\item{dim_Spat}{An integer vector specifying the spatial dimension values.}

\item{path_File}{A string specifying the directory path where the NetCDF file will be created, for [write_nc_WG3] it is the folder path, for [read_nc_WG3] it should be the file path.}

\item{name_Variable}{A string specifying the variable name (must be one of: "airTemprature_Cel", "solarRadition_MJ", "precipatation_mm", "discharge_m3s", "runoff_mm").}

\item{str_Continent}{A string specifying the continent code (must be one of: "eu", "af", "as", "au", "na", "sa").}

\item{suffix_File}{A string to be appended to the filename (typically indicating time period or version).}
}
\value{
None (creates a NetCDF file on disk) / Matrix

A numeric matrix containing the data from the NetCDF file, with rows representing time steps and columns representing spatial units.

A numeric vector containing the data from the NetCDF file, with rows representing time steps and columns representing spatial units.
}
\description{
Write data to a NetCDF file in WaterGAP3 format

Read data from a NetCDF file in WaterGAP3 format
}
\examples{
\dontrun{
data_matrix <- matrix(rnorm(100), nrow=10, ncol=10)
write_nc_WG3(data_matrix, 1:10, 1:10, "/path/to/save", "airTemprature_Cel", "eu", "v1")
}
\dontrun{
data <- read_nc_WG3("/path/to/file.nc", "airTemprature_Cel")
}
\dontrun{
data <- read_nc_dim_WG3("/path/to/file.nc", "time")
}
}
