% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_inflow_cells}
\alias{get_inflow_cells}
\alias{get_inflow_lastcell}
\alias{get_routing_info}
\alias{get_cell_in_basin}
\alias{get_inter_basin}
\alias{get_new_outflow}
\alias{get_cali_step}
\alias{get_upstream_cali_cell}
\title{Get Inflow Cells}
\usage{
get_inflow_cells(int_Outflow)

get_inflow_lastcell(int_Outflow)

get_routing_info(int_Outflow)

get_cell_in_basin(lst_Inflow_Cell, int_OutLet, int_TestCell)

get_inter_basin(lst_Inflow_Cell, int_OutLet, int_UpstreamCell)

get_new_outflow(int_Outflow_Ori, int_CellNew)

get_cali_step(lst_Inflow_Cell, int_CaliCell)

get_upstream_cali_cell(lst_Inflow_Cell, int_CaliCell)
}
\arguments{
\item{int_Outflow}{(vector of int) The cell number of the next cell. The cell number must range from 1 to the length of the cells.
If the cell has no outflow, the number should be set to itself.}

\item{lst_Inflow_Cell}{A list of integer vectors, where each vector contains the cells that flow into the respective cell.}

\item{int_OutLet}{An integer representing the outlet cell (1-based index).}

\item{int_TestCell}{An integer vector, cells to test.}

\item{int_UpstreamCell}{An integer vector containing the upstream cells to find the upstream basin.}

\item{int_Outflow_Ori}{An integer vector representing the original outflow indices (1-based).}

\item{int_CellNew}{An integer vector representing the cells within the new basin.}

\item{int_CaliCell}{An integer vector of calibration cells.}
}
\value{
A list where each element is an IntegerVector containing the inflow cells for each cell.

A NumericMatrix where each row corresponds to a cell, and each column represents an inflow cell.

A list containing the step cells and last cell matrices.

An integer vector of cells in the intersection of the station cells and the basin.

An integer vector representing the new upstream basin, which includes the upstream cells and the set difference of the basin cells.
This function identifies the upstream basin of a given outlet cell by first finding the intersection of the upstream cells
with the cells that flow into the outlet. It then computes the set difference between the upstream basin and the outlet basin.

An integer vector of the new outflow indices adjusted for the sub-basin.

A list of integer vectors (`lst_Step_Cali`), where each element represents calibration cells at a specific step.

A list of integer vectors (`lst_Step_Cali`), where each element represents calibration cells at a specific step.
}
\description{
This function calculates inflow cells based on the outflow vector.

This function creates a matrix of inflow cells for each cell based on the outflow vector.

This function returns a list of step cells and the corresponding last cell matrices.
}
