% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{run_WaterGAP3}
\alias{run_WaterGAP3}
\alias{run_WaterGAP3_N}
\title{run_WaterGAP3}
\usage{
run_WaterGAP3_N(
  name_Region,
  mark_Time,
  n_Time,
  n_Spat,
  path_MeteoInput,
  path_HydroParam,
  path_InitialState = "UNKNOW",
  path_Boundary = "UNKNOW",
  path_VariExport = "NonExport",
  path_FinalState = "NonExport"
)
}
\arguments{
\item{name_Region}{A short identifier of continents: "eu", "af, "as", "au", "na", "sa".}

\item{mark_Time}{A short identifier of time.}

\item{path_MeteoInput}{Directory path containing meteorological input files.}

\item{path_HydroParam}{Directory path containing hydrological parameter files.}

\item{path_InitialState}{Directory path containing initial state files. If "UNKNOW", default initial states will be used.}

\item{path_Boundary}{Directory path containing boundary condition files. If "UNKNOW", default boundary conditions will be used.}

\item{path_VariExport}{Directory path where model output variables (e.g., soil moisture, runoff) will be saved. If "NonExport", no variables will be exported.}

\item{path_FinalState}{Directory path where final state variables will be saved. If "NonExport", final states will not be saved.}
}
\value{
A numeric matrix (CELL_discharge_m3) containing discharge values for each cell and time step.
}
\description{
Run WaterGAP3_N hydrological model

This function runs the WaterGAP3_N hydrological model for a specified region
and time period, returning discharge calculations.
}
