% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{WaterGAP3_H}
\alias{WaterGAP3_H}
\title{WaterGAP3}
\usage{
WaterGAP3_H(
  n_time,
  n_spat,
  atmos_precipitation_mm,
  atmos_temperature_Cel,
  atmos_solarRadiat_MJ,
  snow_ice_mm,
  land_impermeableFrac_1,
  soil_water_mm,
  soil_capacity_mm,
  soil_potentialPercola_mm,
  ground_water_mm,
  ground_capacity_mm,
  basin_landArea_km2,
  lake_cellNumber_int,
  lake_water_m3,
  lake_area_km2,
  lake_capacity_m3,
  river_water_m3,
  river_length_km,
  river_velocity_km,
  basin_cellNumberStep_int,
  basin_inflowCellNumberStep_int,
  param_atmos_thr_Ts,
  param_snow_fac_f,
  param_snow_fac_Tmelt,
  param_evatrans_tur_k,
  param_evatrans_ubc_gamma,
  param_evatrans_vic_gamma,
  param_infilt_hbv_beta,
  param_percola_arn_k,
  param_percola_arn_thresh,
  param_baseflow_grf_gamma,
  param_lake_acp_storeFactor,
  param_lake_acp_gamma
)
}
\arguments{
\item{n_time}{(int) number of time steps}

\item{n_spat}{(int) number of spatial units}

\item{atmos_precipitation_mm}{(mm/m2/TS) precipitaion volum}

\item{atmos_temperature_Cel}{(Cel) the average air temperature in the time phase}

\item{atmos_solarRadiat_MJ}{(MJ/m2/TS) the solar radiation that actually reaches the earths surface}

\item{snow_ice_mm}{(mm/m2) water equivalent of **ice** in snowpack}

\item{land_impermeableFrac_1}{<0, 1> the maximum impermeable fraction when th soil is fully saturated}

\item{soil_water_mm}{(mm/m2) water volume in `soilLy`}

\item{soil_capacity_mm}{(mm/m2) average soil Capacity (maximal storage capacity)}

\item{soil_potentialPercola_mm}{<0.01, 7> (mm/m2/TS) **potential** percolation}

\item{ground_water_mm}{(mm/m2/TS) water volume in `groundLy`}

\item{ground_capacity_mm}{(mm/m2) water storage capacity in `groundLy`}

\item{basin_landArea_km2}{(km2) land area of the basin}

\item{lake_cellNumber_int}{(int) lake cell number}

\item{lake_water_m3}{(m3) lake water volume}

\item{lake_area_km2}{(km2) area of the lake}

\item{lake_capacity_m3}{(m3) capacity of the lake}

\item{river_water_m3}{(m3) river water volume}

\item{river_length_km}{(km) river length}

\item{river_velocity_km}{(km/TS) river velocity}

\item{basin_cellNumberStep_int}{(int) step number of basin cells}

\item{basin_inflowCellNumberStep_int}{(int) step number of inflow cells in the basin}

\item{param_atmos_thr_Ts}{<-1, 3> (Cel) threshold air temperature that snow, parameter for [atmosSnow_ThresholdT()]}

\item{param_snow_fac_f}{<0.05, 2> (mm/m2/h/Cel) potential melt volum per Cel per hour parameter for [snowMelt_Factor()]}

\item{param_snow_fac_Tmelt}{<0, 3> (Cel) snow melt temperature parameter for [snowMelt_Factor()]}

\item{param_evatrans_tur_k}{<0.6, 1> parameter for [evatransPotential_TurcWendling()], higher value when closer to the sea}

\item{param_evatrans_ubc_gamma}{<0.5, 2> parameter for [evatransActual_UBC()]}

\item{param_evatrans_vic_gamma}{<0.2, 5> parameter for [evatransActual_VIC()]}

\item{param_infilt_hbv_beta}{<0.001, 5> parameters for [infilt_HBV()]}

\item{param_percola_arn_k}{<0.1, 1> exponential parameter for [percola_ThreshPow()]}

\item{param_percola_arn_thresh}{<0.1, 0.9> coefficient parameter for [percola_ThreshPow()]}

\item{param_baseflow_grf_gamma}{<2, 7> exponential parameter for [baseflow_GR4Jfix()]}

\item{param_lake_acp_storeFactor}{<uknow> parameter for [lake_AcceptPow()],}

\item{param_lake_acp_gamma}{<uknow> parameter for [lake_AcceptPow()],}
}
\value{
streamflow m3
}
\description{
WaterGAP3
}
