% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{WaterGAP3_H}
\alias{WaterGAP3_H}
\title{WaterGAP3}
\usage{
WaterGAP3_H(
  n_time,
  n_spat,
  ATMOS_precipitation_mm,
  ATMOS_temperature_Cel,
  ATMOS_potentialEvatrans_mm,
  SNOW_ice_mm,
  LAND_builtRatio_1,
  LAND_interceptWater_mm,
  LAND_interceptCapacity_mm,
  SOIL_water_mm,
  SOIL_capacity_mm,
  SOIL_potentialPercola_mm,
  GROUND_water_mm,
  RIVER_water_m3,
  RIVER_length_km,
  RIVER_velocity_km,
  CELL_landArea_km2,
  CELL_cellNumberStep_int,
  CELL_inflowCellNumberStep_int,
  param_ATMOS_thr_Ts,
  param_SNOW_fac_f,
  param_SNOW_fac_Tmelt,
  param_EVATRANS_sup_k,
  param_EVATRANS_sup_gamma,
  param_EVATRANS_wat_petmax,
  param_INFILT_hbv_beta,
  param_PERCOLA_arn_k,
  param_PERCOLA_arn_thresh,
  param_BASEFLOW_sur_k,
  if_allVariExport = FALSE
)
}
\arguments{
\item{n_time}{(int) number of time steps}

\item{n_spat}{(int) number of spatial units}

\item{ATMOS_precipitation_mm}{(mm/m2/TS) precipitaion volum}

\item{ATMOS_temperature_Cel}{(Cel) the average air temperature in the time phase}

\item{ATMOS_potentialEvatrans_mm}{(mm/m2/TS) **potential / reference** evapotranspiration}

\item{SNOW_ice_mm}{(mm/m2) water equivalent of **ice** in snowpack}

\item{LAND_interceptWater_mm}{(mm/m2) initial water volume that can be **intercepted**}

\item{LAND_interceptCapacity_mm}{(mm/m2) average intercept Capacity (maximal storage capacity)}

\item{SOIL_water_mm}{(mm/m2) water volume in `soilLy`}

\item{SOIL_capacity_mm}{(mm/m2) average soil Capacity (maximal storage capacity)}

\item{SOIL_potentialPercola_mm}{<0.01, 7> (mm/m2/TS) **potential** percolation}

\item{GROUND_water_mm}{(mm/m2/TS) water volume in `groundLy`}

\item{RIVER_water_m3}{(m3) river water volume}

\item{RIVER_length_km}{(km) river length}

\item{RIVER_velocity_km}{(km/TS) river velocity}

\item{CELL_landArea_km2}{(km2) land area of the basin}

\item{CELL_cellNumberStep_int}{(int) step number of basin cells}

\item{CELL_inflowCellNumberStep_int}{(int) step number of inflow cells in the basin}

\item{param_ATMOS_thr_Ts}{<-1, 3> (Cel) threshold air temperature that snow, parameter for [atmosSnow_ThresholdT()]}

\item{param_SNOW_fac_f}{<0.05, 2> (mm/m2/h/Cel) potential melt volum per Cel per hour parameter for [snowMelt_Factor()]}

\item{param_SNOW_fac_Tmelt}{<0, 3> (Cel) snow melt temperature parameter for [snowMelt_Factor()]}

\item{param_EVATRANS_sup_k}{<0.1, 1> parameter for [evatransActual_SupplyPow()], ratio of this method}

\item{param_EVATRANS_sup_gamma}{<.1, 5> parameter for [evatransActual_SupplyPow()], exponent of this method}

\item{param_EVATRANS_wat_petmax}{<10, 20> parameter for [evatransActual_WaterGAP()], 10 for arid area, 20 for humid area}

\item{param_INFILT_hbv_beta}{<0.001, 5> parameters for [infilt_HBV()]}

\item{param_PERCOLA_arn_k}{<0.1, 1> exponential parameter for [percola_ThreshPow()]}

\item{param_PERCOLA_arn_thresh}{<0.1, 0.9> coefficient parameter for [percola_ThreshPow()]}

\item{param_BASEFLOW_sur_k}{<0.01, 1> coefficient parameter for [baseflow_SupplyRatio()]}

\item{if_allVariExport}{(bool) whether export all the state variables (default in false)}
}
\value{
streamflow m3
}
\description{
WaterGAP3
}
